Start moving OSD-specific code into its own file
